requirements/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
dcmstack==0.9.0
dicom2nifti==2.4.8
dill==0.3.7
easydict==1.10
et-xmlfile==1.1.0
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
joblib==1.3.1
keras==2.14.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
networkx==3.1
nibabel==5.1.0
nilearn==0.10.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandarallel==1.6.5
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.2
pylibjpeg-libjpeg==1.3.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-gdcm==3.0.22
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
shapely==2.0.1
SimpleITK==2.2.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.30
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
xlrd==2.0.1
